Remove type class from core modules
